fix(logic): use deep update to ensure proper value overrides
Instead of manually deep copying parts of the subconfigs, just deep update the config dict
with the entirety of those subconfigs. This prevents issues where a base config value
is not overriden by a project's own config or a config passed in via the `--config` option,
and keeps the function from needing to be updated if more entries need to be updated
in the future.

Fixes: flakeheaven/flakeheaven#118